+++ lib/phoenix_api_101_web/router.ex
@@ -26,9 +26,9 @@ defmodule PhoenixApi101Web.Router do
     resources("/users", UserController, except: [:new, :edit])
     resources("/posts", PostController, except: [:new, :edit])
 
-    post("/login", SessionController, :login)
-    post("/logout", SessionController, :logout)
-    post("/oauth2/token", SessionController, :refresh_token)
+    # post("/login", AuthController, :login)
+    post("/logout", AuthController, :logout)
+    post("/oauth2/token", AuthController, :login)
     # post("/refresh_token", SessionController, :refresh_token)
     # post("/oauth2/token", AuthController, :callback)
     # delete("/logout", AuthController, :delete)
